% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_summary.R
\name{obj_summary}
\alias{obj_summary}
\alias{obj_summary.default}
\alias{obj_summary.POSIXlt}
\alias{obj_summary.list}
\title{Define how to summarize objects that are contents of list cols}
\usage{
obj_summary(x, ...)

\method{obj_summary}{default}(x, ...)

\method{obj_summary}{POSIXlt}(x, ...)

\method{obj_summary}{list}(x, ...)
}
\arguments{
\item{x}{list element to be parsed}

\item{...}{arguments to be passed to internal functions. not used right now}
}
\value{
string that summarizes the list contents
}
\description{
Table columns can have a class of type "list" which can contain multiple
types. This is an option for how to summarize those object types.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method, reports object type and dimensions

\item \code{POSIXlt}: POSIXlt method, reports number of POSIXlt

\item \code{list}: list method, calculates the typesum and size, summarizes
}}

\examples{
list_example <- list(
    val1 = LETTERS,
    val2 = "1.02e23",
    val2 = data.frame(x = 1)
   )

## summarize various types
obj_summary(list_example[[1]])
obj_summary(list_example[[2]])
obj_summary(list_example[[3]])

}
