% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagedtable.R, R/pagedtable_json.R
\name{pagedtable}
\alias{pagedtable}
\alias{pagedtable_json}
\title{main function for pagedtable}
\usage{
pagedtable(x, ..., width = "100\%", height = "auto")

pagedtable_json(
  x,
  ...,
  use_rownames = getOption("pagedtable.rownames.print"),
  pagerows_min = 10,
  pagerows_max = 100,
  pagecols_min = 1,
  pagecols_max = 100,
  shadowDOM = TRUE
)
}
\arguments{
\item{x}{data.frame to show in the pagedtable}

\item{...}{arguments to be passed to \code{\link{format}}}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{use_rownames}{boolean value. should rownames be used?}

\item{pagerows_min}{integer value defining the minimum number of rows to display for
each page of the pagedtable}

\item{pagerows_max}{integer value defining the maximum number of rows to display for
each page of the pagedtable to fill the height of the widget it is in.}

\item{pagecols_min}{integer value defining the minimum number of columns to display for
each page of the pagedtable}

\item{pagecols_max}{integer value defining the maximum number of columns to display for
each page of the pagedtable to fill the width of the widget it is in.}

\item{shadowDOM}{boolean value, should the shadowDOM be used? defaults to TRUE.}
}
\value{
pagedtable htmlwidget

list object with three entries: data, columns, and options
}
\description{
pagedtable accepts a json that defines the behavior of pagedtable
with the data including: nrows, column names, labels, etc. This function
accepts the data.frame to print and generates the json for paged table
}
\examples{
if(interactive()){
  pagedtable(mtcars)
}

pagedtable_input <- pagedtable_json(mtcars)

}
