% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagedtable_col.R
\name{pagedtable_col}
\alias{pagedtable_col}
\title{Set PagedTable behavior for columns}
\usage{
pagedtable_col(x, name = substitute(x), ...)
}
\arguments{
\item{x}{vector to be parsed}

\item{...}{arguments to be passed to \code{\link{format}}}
}
\value{
list containing two elements, the formatted column, and a series of options
 as defined by pagedtable.js that can change the behavior of the column.
}
\description{
Table column behavior may need to be defined for new/custom s3 vector types.
This s3 method allows for that to be defined!
}
