% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagedtable.R
\name{pagedtable}
\alias{pagedtable}
\title{main function for pagedtable}
\usage{
pagedtable(
  x,
  ...,
  width = "100\%",
  height = "400px",
  pagerows = 10,
  shadowDOM = TRUE
)
}
\arguments{
\item{x}{data.frame to be passed to pagedtable}

\item{...}{arguments to be passed}

\item{pagerows}{rows to display}

\item{shadowDOM}{to or not use shadowdom}
}
\description{
main function for pagedtable
}
