% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagedtable.R
\name{pagedtable}
\alias{pagedtable}
\title{main function for pagedtable}
\usage{
pagedtable(
  x,
  ...,
  width = "100\%",
  height = "auto",
  use_rownames = getOption("pagedtable.rownames.print"),
  pagerows = 10,
  shadowDOM = TRUE
)
}
\arguments{
\item{x}{data.frame to be passed to pagedtable}

\item{...}{arguments to be passed to \code{\link{format}}}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{use_rownames}{should rownames be displayed}

\item{pagerows}{rows to display}

\item{shadowDOM}{to or not use shadowDOM. Default to TRUE}
}
\value{
pagedtable htmlwidget
}
\description{
main function for pagedtable
}
\examples{
if(interactive()){
  pagedtable(mtcars)
}

}
