% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagedtable_col.R
\name{pagedtable_col}
\alias{pagedtable_col}
\alias{pagedtable_col.default}
\alias{pagedtable_col.list}
\title{Set pagedtable behavior for columns}
\usage{
pagedtable_col(x, name = substitute(x), ...)

\method{pagedtable_col}{default}(x, name = substitute(x), ...)

\method{pagedtable_col}{list}(x, name = substitute(x), ...)
}
\arguments{
\item{x}{vector to be parsed}

\item{...}{arguments to be passed to \code{\link{format}}}
}
\value{
list containing two elements, the formatted column, and a series of options
 as defined by pagedtable.js that can change the behavior of the column.
}
\description{
Table column behavior may need to be defined for new/custom s3 vector types.
This s3 method allows for that to be defined!
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method for printing column content

\item \code{list}: list method for printing out list-cols
}}

\examples{

## normal vectors
vector_columns <- pagedtable_col(LETTERS)

## cases like list_col's that are in tibbles
my_list_col <- list(list(1234),"myvalue")
list_columns <- pagedtable_col(my_list_col)

}
